Moved world to scene folder
