+++ src/engine/scene/world.ts
@@ -1,7 +1,7 @@
-import { getContextDataString } from './render-loop'
-import { Position, subtractPositions } from './scene/positionable'
-import Renderable from './scene/renderable'
-import Updatable from './scene/updatable'
+import { getContextDataString } from '../render-loop'
+import { Position, subtractPositions } from './positionable'
+import Renderable from './renderable'
+import Updatable from './updatable'
 
 type Actor = {
     id: string
@@ -16,10 +16,10 @@ export default interface World extends Renderable, Updatable {
     actors: Actor[]
 }
 
-export const defaultWorldPosition: Position = {
+export const defaultWorldPosition = (): Position => ({
     x: 0,
     y: 0
-}
+})
 
 export function renderWorld(world: World, screenOrigin: Position, context: CanvasRenderingContext2D) {
     // console.log(`Rendering world ${getContextDataString(context)}`)
@@ -33,18 +33,17 @@ export function renderWorld(world: World, screenOrigin: Position, context: Canva
     context.strokeRect(0, 0, width, height)
     context.restore()
 
-    const render = (rendering: Renderable) => {
+    for (const actor of world.actors) {
+        if (!actor.rendering) continue
         // console.log(`Starting rendering ${rendering.id} ${getContextDataString(context)}`)
 
         context.save()
-        const renderPosition = subtractPositions(rendering.position, screenOrigin) 
+        const rendering = actor.rendering
+        const renderPosition = subtractPositions(rendering.position, screenOrigin)
         context.translate(renderPosition.x, renderPosition.y)
         rendering.render(context)
         context.restore()
     }
-
-    for (const actor of world.actors) 
-        if (actor.rendering) render(actor.rendering)
 }
 
 export function updateWorld(world: World, deltaTime: number) {
